created analogous color function and tests
